refactor: migrate appError to services

+++ src/modules/answers/services/DeleteAnswerService.ts
@@ -1,6 +1,7 @@
 /* eslint-disable camelcase */
-/* eslint-disable class-methods-use-this */
+import 'reflect-metadata';
 import IAnswersRepository from '@modules/answers/repositories/IAnswersRepository';
+import AppError from '@shared/errors/AppError';
 import { inject, injectable } from 'tsyringe';
 
 @injectable()
@@ -11,9 +12,11 @@ class DeleteAnswerService {
   ) {}
 
   async execute(answer_id:string) {
-    const result = await this.answersRepository.delete(answer_id);
-
-    return result;
+    const answer = await this.answersRepository.findById(answer_id);
+    if (!answer) {
+      throw new AppError('Answer not Found', 404);
+    }
+    await this.answersRepository.delete(answer_id);
   }
 }
